requirements.txt
python-dotenv
openai
telethon
Pillow
requests
